Migrate 'Preferences' and 'Profiles' to own module.
